Remove unreliable test.
The original comment in the "Unstable chains produce different output"
test was right about the version dependence of hash randomization, but
wrong about what version number it required.

In addition, it's possible to disable hash randomization in Perl
versions that would otherwise support it. So rather than create a
convoluted test to see if we have randomized hashes or not, just skip
the test altogether.